Add functionality for templating variables
A new templating functionality has been added. This allows the user to insert variables in a text-based template using an automated text editor UI. This update involves an addition of an 'extension-variables' component, which manages the rendering of variable suggestions and updating of props. The main 'Editor' component has also been updated to utilize this new feature. Certain test files and components associated with the previous 'Annotation' functionality have been removed due to redundancy with this update. The package.json has been updated with necessary dependencies to make this feature work.